Internal: Changed Buffer and Lines to use &str instead of String
This allows for less allocations and therefore a reduced memory usage
This is important mostly for large files.
The memory usage is indeed reduced in this version but the program
still allocates much more than the size of the actual file while
the actual less command seems to not use any memory at all.
The less command does seem to lag on certain commands on large files
such as quitting when going to the end of a large file. Maybe less
dynamically collects information about the file which is then read
when needed? I would need to study the source of less to see what
it does. I must be doing something smarter than what I am doing here.